chore(sidebar): reorganise sidebar utilities into lib/sidebar/
- Create web/src/lib/sidebar/{constants,utils,svc,hooks}.ts — canonical
  home for sidebar constants, utility fns, API orchestration, and the
  useChatSearchOptimistic hook
- Move formatDisplayTime to @opal/time with full JSDoc; delete
  chatSearchUtils.ts
- Delete useChatSearchOptimistic.ts, sidebarUtils.ts, constants.ts (all
  superseded by the lib/sidebar modules)
- Delete SidebarBody.tsx; craft/SideBar.tsx now uses SidebarLayouts.*
  directly (Root + Header + Body + Footer)
- Drop DEFAULT_PERSONA_ID from sidebar constants (duplicate); inline 0
  at callsites
- Fix all relative imports to absolute @/ paths
