+++ web/src/app/craft/components/SideBar.tsx
@@ -18,8 +18,10 @@ import {
   useSidebarState,
 } from "@opal/layouts";
 import RefreshText from "@/refresh-components/texts/Text";
-import SidebarWrapper from "@/sections/sidebar/SidebarWrapper";
-import SidebarBody from "@/sections/sidebar/SidebarBody";
+import {
+  renderAppLogo,
+  useShowLogoWhenFolded,
+} from "@/sections/sidebar/SidebarWrapper";
 import AccountPopover from "@/sections/sidebar/AccountPopover";
 import { Popover, PopoverMenu } from "@opal/components";
 import IconButton from "@/refresh-components/buttons/IconButton";
@@ -449,20 +451,22 @@ const MemoizedBuildSidebarInner = memo(() => {
     [folded, backToChatButton]
   );
 
+  const showLogoWhenFolded = useShowLogoWhenFolded();
+
   return (
-    <SidebarWrapper foldable>
-      <SidebarBody
-        pinnedContent={
-          <div className="flex flex-col gap-0.5">
-            {newBuildButton}
-            {scheduledTasksPanel}
-            {skillsPanel}
-            {appsTab}
-          </div>
-        }
-        footer={footer}
-        scrollKey="build-sidebar"
+    <SidebarLayouts.Root foldable>
+      <SidebarLayouts.Header
+        logo={renderAppLogo}
+        showLogoWhenFolded={showLogoWhenFolded}
       >
+        <div className="flex flex-col gap-0.5">
+          {newBuildButton}
+          {scheduledTasksPanel}
+          {skillsPanel}
+          {appsTab}
+        </div>
+      </SidebarLayouts.Header>
+      <SidebarLayouts.Body scrollKey="build-sidebar">
         {!folded && (
           <>
             <SidebarLayouts.Section title={sessionsTitle} />
@@ -498,8 +502,9 @@ const MemoizedBuildSidebarInner = memo(() => {
             )}
           </>
         )}
-      </SidebarBody>
-    </SidebarWrapper>
+      </SidebarLayouts.Body>
+      <SidebarLayouts.Footer>{footer}</SidebarLayouts.Footer>
+    </SidebarLayouts.Root>
   );
 });
