+++ web/src/lib/sidebar/constants.ts
@@ -7,5 +7,3 @@ export const DRAG_TYPES = {
 export const LOCAL_STORAGE_KEYS = {
   HIDE_MOVE_CUSTOM_AGENT_MODAL: "onyx:hideMoveCustomAgentModal",
 } as const;
-
-export const DEFAULT_PERSONA_ID = 0;
