+++ web/src/lib/sidebar/hooks.ts
@@ -31,8 +31,6 @@ interface UseChatSearchOptimisticResult {
 const PAGE_SIZE = 20;
 const DEBOUNCE_MS = 300;
 
-// --- Helper Functions ---
-
 function transformApiResponse(response: ChatSearchResponse): FilterableChat[] {
   const chats: FilterableChat[] = [];
   for (const group of response.groups) {
@@ -58,28 +56,20 @@ function filterLocalSessions(
   return sessions.filter((chat) => chat.label.toLowerCase().includes(term));
 }
 
-// --- Hook ---
-
 export default function useChatSearchOptimistic(
   options: UseChatSearchOptimisticOptions
 ): UseChatSearchOptimisticResult {
   const { searchQuery, enabled = true } = options;
 
-  // Debounced search query for API calls
   const [debouncedQuery, setDebouncedQuery] = useState(searchQuery);
-
-  // Ref for infinite scroll sentinel
   const sentinelRef = useRef<HTMLDivElement | null>(null);
 
-  // 1. Get already-cached data from existing hooks
   const { chatSessions } = useChatSessions();
   const { projects } = useProjects();
 
-  // 2. Build combined fallback data (instant display)
   const fallbackSessions = useMemo<FilterableChat[]>(() => {
     const chatMap = new Map<string, FilterableChat>();
 
-    // Add regular chats from useChatSessions
     for (const chat of chatSessions) {
       chatMap.set(chat.id, {
         id: chat.id,
@@ -88,7 +78,6 @@ export default function useChatSearchOptimistic(
       });
     }
 
-    // Add project chats from useProjects
     for (const project of projects) {
       for (const chat of project.chat_sessions) {
         chatMap.set(chat.id, {
@@ -99,25 +88,19 @@ export default function useChatSearchOptimistic(
       }
     }
 
-    // Sort by most recent
     return Array.from(chatMap.values()).sort(
       (a, b) => new Date(b.time).getTime() - new Date(a.time).getTime()
     );
   }, [chatSessions, projects]);
 
-  // Debounce the search query
   useEffect(() => {
     const timer = setTimeout(() => setDebouncedQuery(searchQuery), DEBOUNCE_MS);
     return () => clearTimeout(timer);
   }, [searchQuery]);
 
-  // 3. SWR key generator for infinite scroll
   const getKey = useCallback(
     (pageIndex: number, previousPageData: ChatSearchResponse | null) => {
-      // Don't fetch if not enabled
       if (!enabled) return null;
-
-      // Reached the end
       if (previousPageData && !previousPageData.has_more) return null;
 
       const page = pageIndex + 1;
@@ -134,7 +117,6 @@ export default function useChatSearchOptimistic(
     [enabled, debouncedQuery]
   );
 
-  // 4. Use SWR for paginated data (replaces fallback after fetch)
   const { data, size, setSize, isValidating } =
     useSWRInfinite<ChatSearchResponse>(getKey, errorHandlingFetcher, {
       revalidateOnFocus: false,
@@ -143,7 +125,6 @@ export default function useChatSearchOptimistic(
       persistSize: true,
     });
 
-  // Transform SWR data to FilterableChat[]
   const swrResults = useMemo<FilterableChat[]>(() => {
     if (!data || data.length === 0) return [];
 
@@ -152,7 +133,6 @@ export default function useChatSearchOptimistic(
       allChats.push(...transformApiResponse(page));
     }
 
-    // Deduplicate by id (keep first occurrence)
     const seen = new Set<string>();
     return allChats.filter((chat) => {
       if (seen.has(chat.id)) return false;
@@ -161,22 +141,17 @@ export default function useChatSearchOptimistic(
     });
   }, [data]);
 
-  // Determine if we have more pages
   const hasMore = useMemo(() => {
     if (!data || data.length === 0) return true;
     const lastPage = data[data.length - 1];
     return lastPage?.has_more ?? false;
   }, [data]);
 
-  // 5. Return fallback if no SWR data yet, otherwise return SWR data
   const results = useMemo<FilterableChat[]>(() => {
-    // If SWR has data, use it (paginated, searchable)
     if (swrResults.length > 0) {
       return swrResults;
     }
 
-    // Otherwise use fallback (already-cached data)
-    // Apply local filtering if there's a search query
     if (searchQuery.trim()) {
       return filterLocalSessions(fallbackSessions, searchQuery);
     }
@@ -184,11 +159,9 @@ export default function useChatSearchOptimistic(
     return fallbackSessions;
   }, [swrResults, fallbackSessions, searchQuery]);
 
-  // Loading states
   const isSearching = isValidating && size === 1;
   const isLoadingMore = isValidating && size > 1;
 
-  // Fetch more results for infinite scroll
   const fetchMore = useCallback(async () => {
     if (!enabled || isValidating || !hasMore) {
       return;
@@ -196,7 +169,6 @@ export default function useChatSearchOptimistic(
     await setSize(size + 1);
   }, [enabled, isValidating, hasMore, setSize, size]);
 
-  // IntersectionObserver for infinite scroll
   useEffect(() => {
     const sentinel = sentinelRef.current;
     if (!sentinel || !enabled) return;
@@ -208,18 +180,11 @@ export default function useChatSearchOptimistic(
           fetchMore();
         }
       },
-      {
-        root: null,
-        rootMargin: "100px",
-        threshold: 0,
-      }
+      { root: null, rootMargin: "100px", threshold: 0 }
     );
 
     observer.observe(sentinel);
-
-    return () => {
-      observer.disconnect();
-    };
+    return () => observer.disconnect();
   }, [enabled, hasMore, isValidating, fetchMore]);
 
   return {
