+++ web/src/sections/sidebar/ChatSearchCommandMenu.tsx
@@ -9,14 +9,12 @@ import CommandMenu, {
 import { useProjects } from "@/lib/hooks/useProjects";
 import { useCreateModal } from "@/refresh-components/contexts/ModalContext";
 import CreateProjectModal from "@/sections/modals/CreateProjectModal";
-import {
-  formatDisplayTime,
-  highlightMatch,
-} from "@/sections/sidebar/chatSearchUtils";
+import { formatDisplayTime } from "@opal/time";
+import { highlightMatch } from "@/lib/sidebar/utils";
 import { useSettingsContext } from "@/providers/SettingsProvider";
 import { useCurrentAgent } from "@/lib/agents/hooks";
 import Text from "@/refresh-components/texts/Text";
-import useChatSearchOptimistic from "./useChatSearchOptimistic";
+import useChatSearchOptimistic from "@/lib/sidebar/hooks";
 import {
   SvgEditBig,
   SvgFolder,
