+++ web/src/sections/sidebar/ProjectFolderButton.tsx
@@ -10,7 +10,7 @@ import ChatButton from "@/sections/sidebar/ChatButton";
 import { useAppRouter } from "@/hooks/appNavigation";
 import { noProp } from "@/lib/utils";
 import { cn } from "@opal/utils";
-import { DRAG_TYPES } from "./constants";
+import { DRAG_TYPES } from "@/lib/sidebar/constants";
 import IconButton from "@/refresh-components/buttons/IconButton";
 import Truncated from "@/refresh-components/texts/Truncated";
 import ButtonRenaming from "@/refresh-components/buttons/ButtonRenaming";
